performance!: improve performance of authContext
By wrapping the functions declared inside the component with useCallback, we prevent the creation of new function references in every render. This significantly improves performance and reduces the risk of overflow due to excessive function calls. Additionally, using the useMemo React hook to wrap our objects ensures that expensive calculations are not performed between renders if the objects have not changed.